feat: implement storage module with delete file functionality
